ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.92.1
